+++ src/slack/conversation.js
@@ -72,21 +72,59 @@ class Conversation {
 
 
   /**
-   * Asks the question and waits for the answer
+   * Asks the question and waits for the answer.
+   * If format provided asks again until enforced replyPattern requirements met.
+   *
+   * @param {(string|Object)} opts - String, or object.
+   * @param {string=} opts.text - Question.
+   * @param {RegExp=} opts.replyPattern - Reply pattern as an instance of RegExp.
+   * @param {(function|Promise)} cb - Callback. Has parameter `response`.
+   *
    * @returns {Promise}
    */
-  ask(question) {
+  ask(opts, cb) {
+    if (!opts) return Promise.reject(new Error('Unknown question object/string.'));
+
+    /**
+     * Opts can be an object or a string.
+     */
+    if (typeof(opts) == 'string') {
+      opts = {
+        text: opts
+      };
+    }
+
+    /**
+     * Add a reply pattern to check if the response fits your needs.
+     */
+    if(opts.replyPattern && !(opts.replyPattern instanceof RegExp)) return Promise.reject(new Error('replyPattern is not valid. It should be an instance of RegExp.'));
+
     let that = this;
 
     debug('Ask question to the user.');
     return this
-      .say(question)
+      .say(opts.text)
       .then(() => {
         return new Promise((resolve) => {
           debug('Wait for a response.');
           that.eventEmitter.once('message', (response) => {
             debug('Response received');
-            resolve(response);
+            if (opts.replyPattern && opts.replyPattern.test(response.text))
+              return resolve(response);
+
+            if (cb) {
+              /**
+               * Check if callback is promisified. If it's promisified, wait for it.
+               */
+              if (typeof cb.then == 'function')
+                return cb(response).then(() => {
+                  return resolve(that.ask(opts, cb));
+                });
+
+              cb(response);
+            }
+
+            resolve(that.ask(opts, cb));
           });
         })
       });
@@ -95,12 +133,18 @@ class Conversation {
 
   /**
    * Asks an array of questions while waiting for the answer of each.
+   *
+   * @param {Object[]} opts - Array of opt objects.
+   * @param {string} opts[].text - Question to be asked.
+   * @param {string} opts[].replyPattern - Reply pattern to be enforced.
+   * @param {Function} opts[].callback - Callback to call upon faulty replies.
+   *
    * @returns {Promise}
    */
-  askSerial(questions) {
+  askSerial(opts) {
     let that = this;
-    return Promise.mapSeries(questions, (question) => {
-      return that.ask(question)
+    return Promise.mapSeries(opts, (opt) => {
+      return that.ask(opt, opt.callback || null)
         .then((response) => {
           return response;
         });